refactor: split js into modules
