+++ index.html
@@ -97,6 +97,6 @@
     <button id="accept-consent">Accept</button>
   </div>
 
-  <script src="assets/js/index.js"></script>
+  <script type="module" src="assets/js/index.js"></script>
 </body>
 </html>
